Adding clear filter PcapDevice

+++ Packet++/header/IcmpLayer.h
@@ -144,7 +144,7 @@ namespace pcpp
 		/** a pointer to the header data */
 		icmp_echo_hdr* header;
 		/** most echo requests/replies contain some payload data. This is the data length */
-		uint8_t dataLength;
+		size_t dataLength;
 		/** most echo requests/replies contain some payload data. This is a pointer to this data */
 		uint8_t* data;
 	} icmp_echo_request;
